+++ test/libsolidity/syntaxTests/license/license_double.sol
@@ -2,4 +2,4 @@
 contract C {}
 // SPDX-License-Identifier: MIT
 // ----
-// ParserError 3716: Multiple SPDX license identifiers found in source file. Use "AND" or "OR" to combine multiple licenses. Please see https://spdx.org for more information.
+// Warning 1878: SPDX license identifier not provided in source file. Before publishing, consider adding a comment containing "SPDX-License-Identifier: <SPDX-License>" to each source file. Use "SPDX-License-Identifier: UNLICENSED" for non-open-source code. Please see https://spdx.org for more information.